SCREAM: allow atm procs to require/update a "group" of fields.
This means the AP does not know the names of the fields it needs/updates,
or even how many they are. It will perform the same operation on all of them,
which must be a "generic" operation (i.e., can be done on a generic field).
Examples are advecting them, or apply a fixer/limiter.

Notable example: dynamics can advect a bunch of fields, but it doesn't need
to know the names of the fields (the tracers), which means it may not be
able to build their identifier (which is needed for regular required/computed
fields). By allowing to request a "group", it can ask for, say, all fields
that have the "ADVECT_ME" label.