Added infrastructure for perturbable fields.
The AtmosphereDriver class now creates a set_perturbable_field()
alongside the existing set_required_field() and
set_computed_field(), containing the union of both but with the
required fields as non-const.  This will allow perturbing the
input values to generate machine learning training sets.

set_perturbable_field_impl() is a virtual function in the
AtmosphereProcess class, and thus any class derived from it
needed and implementation.  That's while so many files were
modified.